pyproject.toml
[tool.poetry]
name = "seq-rec"
version = "0.1.0"
description = ""
authors = ["Quy Dinh <dvquy.13@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.7,<3.11"
tensorflow = "2.9.1"
tensorflow-datasets = "4.6.0"
tensorflow-estimator = "2.9.0"
tensorflow-gcs-config = "2.8.0"
tensorflow-gpu = "2.9.1"
tensorflow-hub = "0.12.0"
tensorflow-io = "0.26.0"
tensorflow-io-gcs-filesystem = "0.26.0"
tensorflow-metadata = "1.8.0"
tensorflow-probability = "0.16.0"
tensorflow-recommenders = "0.6.0"
scann = "*"
google-cloud-bigquery-storage = "*"
fastavro = "*"
google-cloud-aiplatform = "*"
comet-ml = "^3.31.3"
hydra-core = "*"
python-dotenv = "*"
google-api-python-client = "*"
google-auth-httplib2 = "*"
google-auth-oauthlib = "*"
kfp = ">=1.8.9"
google-cloud-pipeline-components = "*"
google-cloud-storage = "*"
cloudpickle = ">=2.0.0"

[tool.poetry.dev-dependencies]

[tool.poetry2conda]
name = "dl-context"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
